+++ internal/old/output/constructor.go
@@ -227,7 +227,6 @@ type Config struct {
 	MQTT               writer.MQTTConfig              `json:"mqtt" yaml:"mqtt"`
 	Nanomsg            writer.NanomsgConfig           `json:"nanomsg" yaml:"nanomsg"`
 	NATS               writer.NATSConfig              `json:"nats" yaml:"nats"`
-	NATSJetStream      NATSJetStreamConfig            `json:"nats_jetstream" yaml:"nats_jetstream"`
 	NATSStream         writer.NATSStreamConfig        `json:"nats_stream" yaml:"nats_stream"`
 	NSQ                writer.NSQConfig               `json:"nsq" yaml:"nsq"`
 	Plugin             interface{}                    `json:"plugin,omitempty" yaml:"plugin,omitempty"`
@@ -287,7 +286,6 @@ func NewConfig() Config {
 		MongoDB:            NewMongoDBConfig(),
 		Nanomsg:            writer.NewNanomsgConfig(),
 		NATS:               writer.NewNATSConfig(),
-		NATSJetStream:      NewNATSJetStreamConfig(),
 		NATSStream:         writer.NewNATSStreamConfig(),
 		NSQ:                writer.NewNSQConfig(),
 		Plugin:             nil,